# public/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "StatusHub API",
    "version": "1.0.0",
    "description": "API REST para monitorización de servicios IT en tiempo real.",
    "contact": {
      "email": "admin@statushub.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Servidor local (Docker)"
    }
  ],
  "components": {
    "securitySchemes": {
      "sanctum": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Token",
        "description": "Introduce el token obtenido en /api/auth/login"
      }
    }
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Iniciar sesión",
        "description": "Autentica al usuario y devuelve un token de acceso.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "admin@statushub.com" },
                  "password": { "type": "string", "example": "password" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login correcto",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer", "example": 1 },
                        "name": { "type": "string", "example": "Admin" },
                        "email": { "type": "string", "example": "admin@statushub.com" },
                        "role": { "type": "string", "example": "admin" }
                      }
                    },
                    "token": { "type": "string", "example": "1|abc123..." }
                  }
                }
              }
            }
          },
          "422": { "description": "Credenciales incorrectas" }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Cerrar sesión",
        "description": "Invalida el token de acceso del usuario autenticado.",
        "security": [{ "sanctum": [] }],
        "responses": {
          "200": {
            "description": "Sesión cerrada correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Sesión cerrada correctamente." }
                  }
                }
              }
            }
          },
          "401": { "description": "No autenticado" }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": ["Autenticación"],
        "summary": "Usuario autenticado",
        "description": "Devuelve los datos del usuario que ha iniciado sesión.",
        "security": [{ "sanctum": [] }],
        "responses": {
          "200": {
            "description": "Datos del usuario autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "integer", "example": 1 },
                    "name": { "type": "string", "example": "Admin" },
                    "email": { "type": "string", "example": "admin@statushub.com" },
                    "role": { "type": "string", "example": "admin" }
                  }
                }
              }
            }
          },
          "401": { "description": "No autenticado" }
        }
      }
    },
...
  }
}